+++ README.md
@@ -118,6 +118,15 @@ for (char alp = 'A'; alp <= 'Z'; alp++) { }
 ### 백준 10431 줄세우기
 - https://www.acmicpc.net/problem/10431
 - 삽입 정렬의 개념을 적용하면 풀 수 있는 문제였는데 삽입 정렬의 아이디어를 떠올리지 못했다. 삽입 정렬 알고리즘에 대해서 다시 복습하자.
+
+### 백준 10989 수 정렬하기 3
+- https://www.acmicpc.net/problem/10989
+- 문자열 index를 사용해서 문자의 갯수를 셌던 아이디어를 수열에도 똑같이 적용할 수 있다. 일반적인 정렬 알고리즘의 최악의 시간복잡도는 O(N^2)이므로 입력되는 수의 개수가 일천만이 넘아간다면 일반적일 정렬 방법으로는 시간 초과가 날 가능성이 높다.
+
+### 백준 3273 두 수의 합
+- https://www.acmicpc.net/problem/3273
+- 시간 제한이 1초고 N의 크기가 백만이라 배열을 정렬하면 무조건 시간 초과가 난다. 백준 10989의 문제처럼 백열의 index를 활용해서 풀 수 있다. `pairValue = X - num[i]`라고 할 때 `exist[pairValue]` 배열에 존재한다면 입력받은 모든 값들을 순회하지 않아도 답을 구할 수 있다. 
+
 ## Part 5. 단기완성 알고리리즘
 ## 완전 탐색(Brute Force)
 ### 백준 15649 N과 M(1) ~ 백준 15652N과 M(4)
